# Improve batch item size estimation