--- a/sdks/python/src/opik/message_processing/batching/sequence_splitter.py
+++ b/sdks/python/src/opik/message_processing/batching/sequence_splitter.py
@@ -1,14 +1,59 @@
-import json
-from typing import List, Optional, TypeVar, Sequence
+import logging
+from typing import List, Optional, TypeVar, Sequence, Any
 from opik import jsonable_encoder
 
 T = TypeVar("T")
 
+LOGGER = logging.getLogger(__name__)
+
 
 def _get_expected_payload_size_MB(item: T) -> float:
     encoded_for_json = jsonable_encoder.encode(item)
-    json_str = json.dumps(encoded_for_json)
-    return len(json_str.encode("utf-8")) / (1024 * 1024)
+    size = _get_json_size(encoded_for_json)
+    return size
+
+
+def _get_json_size(obj: Any) -> Any:
+    """
+    Compute the size of the resulting JSON without actually doing the JSON
+    encoding which is CPU and memory consuming. This assumes that we only
+    receives basic Python objects, strings, boolean, numbers, list and dicts
+    and that the object do not contains any cyclic reference.
+    """
+    try:
+        if isinstance(obj, str):
+            return len(obj.encode("utf-8")) + 2  # "str_content"
+        elif isinstance(obj, (int, float)):
+            return len(str(obj))
+        elif isinstance(obj, type(None)):
+            # "null"
+            return 6
+        elif isinstance(obj, dict):
+            size = 2  # {obj}
+            allowed_keys = set(obj.keys())
+            for key, value in obj.items():
+                if key in allowed_keys:
+                    encoded_key = _get_json_size(key)
+                    encoded_value = _get_json_size(value)
+                    size += encoded_key + encoded_value + 1  # key:value
+            return size
+        elif isinstance(obj, list):
+            size = 2  # [obj]
+            for item in obj:
+                size += _get_json_size(item) + 1  # ,
+            return size - 1  # Remove the last trailing comma
+        elif isinstance(obj, bool):
+            return len(str(obj))
+        else:
+            LOGGER.debug(
+                "Unexpected object seen during JSON size estimation %r", type(obj)
+            )
+            return len(str(obj))
+
+    except Exception:
+        LOGGER.debug("Failed to compute object size.", exc_info=True)
+        # Return a value that will cause the span to be in its own batch to be on the safe side
+        return float("inf")
 
 
 def split_into_batches(