--- a/sdks/python/tests/unit/message_processing/batching/test_sequence_splitter.py
+++ b/sdks/python/tests/unit/message_processing/batching/test_sequence_splitter.py
@@ -1,4 +1,8 @@
+import json
 import dataclasses
+
+import pytest
+
 from opik.message_processing.batching import sequence_splitter
 
 
@@ -75,3 +79,60 @@ def test_split_list_into_batches__multiple_large_objects():
         [ONE_MEGABYTE_OBJECT_B],
         [ONE_MEGABYTE_OBJECT_C],
     ]
+
+
+@pytest.mark.parametrize(
+    "test_input",
+    [
+        ("string"),  # Basic string
+        (""),  # Empty string
+        ("Unicode👋🏼"),  # String with non-UTF8 characters
+        (123.76),  # Float
+        (123456789),  # Integer
+        (True),  # Boolean
+        (False),  # Boolean
+        ([1, 2, 3]),  # List
+        ({"key": "value"}),  # Dictionary
+        (None),  # null
+    ],
+)
+def test_get_json_size_basic(test_input):
+    expected = len(
+        json.dumps(test_input, separators=(",", ":"), ensure_ascii=False).encode(
+            "utf-8"
+        )
+    )
+    assert sequence_splitter._get_json_size(test_input) == expected
+
+
+def test_get_json_size_complex_nested():
+    test_input = {
+        "string": "simple value",
+        "number": 42,
+        "boolean": True,
+        "null": None,
+        "nested_dict": {
+            "level1": {
+                "level2": {
+                    "level3": "deep value",
+                    "numbers": [1, 2, 3],
+                    "mixed_list": [True, "string", 42.0, None],
+                },
+                "sibling2": {
+                    "data": [{"id": 1, "name": "item1"}, {"id": 2, "name": "item2"}]
+                },
+            }
+        },
+        "array_of_arrays": [[1, 2, 3], ["a", "b", "c"], [{"x": 1}, {"y": 2}]],
+        "mixed_nesting": {
+            "lists": [[1, 2], [3, 4]],
+            "dicts": {"a": {"b": {"c": "value"}}, "x": {"y": {"z": 100}}},
+        },
+    }
+
+    expected = len(
+        json.dumps(test_input, separators=(",", ":"), ensure_ascii=False).encode(
+            "utf-8"
+        )
+    )
+    assert sequence_splitter._get_json_size(test_input) == expected