New algorithm is as fast but takes minimal memory versus the old algorithm

## Details

I tested against a 30 MB spans batch.

In term of speed, this is equivalent if not slightly faster:

main branch:
```
654 ms ± 3.67 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

this branch:
```
620 ms ± 22.2 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

In term of memory, I measure a spike of 20MB in the main branch and couldn't measure any memory spike in the new branch (versus reading just the data in memory).